backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
aio-pika==9.4.1
pydantic==2.7.4
pydantic-settings==2.3.3
python-dotenv==1.0.1
email-validator==2.1.1
